Rename registerNewNetwork
